requirements.txt
python-telegram-bot==21.0.1
Pillow==11.0.0
pdf2image==1.16.3
PyPDF2==3.0.1
reportlab==4.0.7  # For PDF generation

